products

+++ ManagementSystem/ext/webui/admin.py
@@ -7,14 +7,18 @@ from flask import *
 from flask_login import *
 from flask_toastr import *
 
-from ManagementSystem.ext.database.attendances import delete_attendance, add_attendance, update_attendance
 from ManagementSystem.ext.crypt import encrypt_id_with_no_digits
-from ManagementSystem.ext.database.users import get_users_by_role, get_user_by_id
+from ManagementSystem.ext.database.attendances import delete_attendance, add_attendance, update_attendance
 from ManagementSystem.ext.database.courses import get_courses, add_course, delete_course, update_course, \
     check_course_by_name
 from ManagementSystem.ext.database.maps import get_map_by_name, update_trips
+from ManagementSystem.ext.database.offers import get_offers, delete_offer, add_offer
+from ManagementSystem.ext.database.orders import get_orders
+from ManagementSystem.ext.database.products import get_products, get_product_by_id, add_product, update_product, \
+    delete_product
 from ManagementSystem.ext.database.records import get_records_by_author, get_records_by_type, RecordType, add_record, \
     update_record_news, delete_record
+from ManagementSystem.ext.database.users import get_users_by_role, get_user_by_id
 from ManagementSystem.ext.logistics import auto_redirect, check_session
 from ManagementSystem.ext.models.userModel import Role
 from ManagementSystem.ext.telegram_bot.message import send_news
@@ -109,17 +113,20 @@ def admin_news():
                         types = ['jpeg', 'jpg', 'png']
                         if img_type in types:
                             resp_status, data = encrypt_id_with_no_digits(str(inserted_record.inserted_id))
-                            record_id = data
-                            directory = 'storage/records/'
-                            files = os.listdir(directory)
-                            for file in files:
-                                if record_id == file.split('.')[0]:
-                                    filename = file
-                                    break
-                            if filename != '':
-                                os.remove(directory + filename)
-                            filename = record_id + '.' + img_type
-                            image.save(directory + filename)
+                            if resp_status:
+                                record_id = data
+                                directory = 'storage/records/'
+                                files = os.listdir(directory)
+                                for file in files:
+                                    if record_id == file.split('.')[0]:
+                                        filename = file
+                                        break
+                                if filename != '':
+                                    os.remove(directory + filename)
+                                filename = record_id + '.' + img_type
+                                image.save(directory + filename)
+                            else:
+                                flash('Не удалось обработать данные.', 'warning')
                         else:
                             flash('Недопустимый тип файла.', 'warning')
                 except Exception as ex:
@@ -170,18 +177,21 @@ def admin_news():
                         types = ['jpeg', 'jpg', 'png']
                         if img_type in types:
                             resp_status, data = encrypt_id_with_no_digits(str(record_id))
-                            record_id = data
-                            filename = ''
-                            directory = 'storage/records/'
-                            files = os.listdir(directory)
-                            for file in files:
-                                if record_id == file.split('.')[0]:
-                                    filename = file
-                                    break
-                            if filename != '':
-                                os.remove(directory + filename)
-                            filename = record_id + '.' + img_type
-                            image.save(directory + filename)
+                            if resp_status:
+                                record_id = data
+                                filename = ''
+                                directory = 'storage/records/'
+                                files = os.listdir(directory)
+                                for file in files:
+                                    if record_id == file.split('.')[0]:
+                                        filename = file
+                                        break
+                                if filename != '':
+                                    os.remove(directory + filename)
+                                filename = record_id + '.' + img_type
+                                image.save(directory + filename)
+                            else:
+                                flash('Не удалось обработать данные.', 'warning')
                         else:
                             flash('Недопустимый тип файла.', 'warning')
                 except Exception as ex:
@@ -192,16 +202,19 @@ def admin_news():
                 rec_id = request.form.get("record_id")
                 try:
                     resp_status, data = encrypt_id_with_no_digits(str(rec_id))
-                    record_id = data
-                    filename = ''
-                    directory = 'storage/records/'
-                    files = os.listdir(directory)
-                    for file in files:
-                        if record_id == file.split('.')[0]:
-                            filename = file
-                            break
-                    if filename != '':
-                        os.remove(directory + filename)
+                    if resp_status:
+                        record_id = data
+                        filename = ''
+                        directory = 'storage/records/'
+                        files = os.listdir(directory)
+                        for file in files:
+                            if record_id == file.split('.')[0]:
+                                filename = file
+                                break
+                        if filename != '':
+                            os.remove(directory + filename)
+                    else:
+                        flash('Не удалось обработать данные.', 'warning')
                 except Exception as ex:
                     logger.error(ex)
 
@@ -400,3 +413,201 @@ def user_attendance(user_id):
         return render_template("error-500.html")
 
     return render_template("admin/courses/user-attendance.html", user=resp.data)
+
+
+# Уровень:              admin/offers
+# База данных:          offers
+# HTML:                 offers
+@admin.route('/offers', methods=['POST', 'GET'])
+@login_required
+def admin_offers():
+    # auto redirect
+    status, url = auto_redirect(ignore_role=Role.ADMIN)
+    if status:
+        return redirect(url)
+    # check session
+    if not check_session():
+        logout_user()
+        return redirect(url_for("view.landing"))
+
+    if request.method == "POST":
+        try:
+            if request.form['btn_offers'] == 'add_offer':
+                name = request.form.get('name')
+                description = request.form.get('description')
+                date = request.form.get('date')
+                start_time = request.form.get('start_time')
+                finish_time = request.form.get('finish_time')
+                reward = request.form.get('reward')
+                if len(start_time) > 0 and len(finish_time) > 0:
+                    dates = date.replace(' ', '').split('-')
+                    start_time = '0' * (2 - len(start_time)) + start_time
+                    finish_time = '0' * (2 - len(finish_time)) + finish_time
+                    start_date = datetime.strptime(dates[0] + ' ' + start_time, '%m/%d/%Y %H:%M')
+                    finish_date = datetime.strptime(dates[1] + ' ' + finish_time, '%m/%d/%Y %H:%M')
+                    add_offer(author=current_user.id, name=name, description=description, start=start_date,
+                              finish=finish_date, reward=reward)
+                    flash('Вы успешно добавили оффер', 'success')
+                else:
+                    flash('Выберите дату и время', 'warning')
+            elif request.form['btn_offers'] == 'delete_offer':
+                offer_id = request.form.get('offer_id')
+                delete_offer(offer_id)
+                flash('Вы успешно удалили оффер', 'success')
+        except Exception as ex:
+            logger.error(ex)
+            flash('Произошла какая-то ошибка', 'error')
+
+    return render_template("admin/other/offers.html", offers=get_offers().data)
+
+
+# Уровень:              admin/products
+# База данных:          products
+# HTML:                 products
+@admin.route('/products', methods=['POST', 'GET'])
+@login_required
+def admin_products():
+    # auto redirect
+    status, url = auto_redirect(ignore_role=Role.ADMIN)
+    if status:
+        return redirect(url)
+    # check session
+    if not check_session():
+        logout_user()
+        return redirect(url_for("view.landing"))
+
+    if request.method == "POST":
+        try:
+            if request.form['btn_products'] == 'add_product':
+                name = request.form.get("name")
+                info = request.form.get("description")
+                price = request.form.get("price")
+
+                inserted_product = add_product(name, info, price)
+                flash('Вы добавили товар.', 'success')
+
+                # image
+                filename = ''
+                try:
+                    image = request.files['image']
+                    img_type = image.filename.split('.')[-1].lower()
+                    if img_type != '':
+                        types = ['jpeg', 'jpg', 'png']
+                        if img_type in types:
+                            resp_status, data = encrypt_id_with_no_digits(str(inserted_product.inserted_id))
+                            if resp_status:
+                                product_id = data
+                                directory = 'storage/products/'
+                                files = os.listdir(directory)
+                                for file in files:
+                                    if product_id == file.split('.')[0]:
+                                        filename = file
+                                        break
+                                if filename != '':
+                                    os.remove(directory + filename)
+                                filename = product_id + '.' + img_type
+                                image.save(directory + filename)
+                            else:
+                                flash('Не удалось обработать данные.', 'warning')
+                        else:
+                            flash('Недопустимый тип файла.', 'warning')
+                except Exception as ex:
+                    flash('Не удалось прикрепить изображение.', 'warning')
+                    logger.error(ex)
+            elif request.form['btn_products'] == 'edit_product':
+                prod_id = request.form.get("product_id")
+                name = request.form.get("name")
+                info = request.form.get("description")
+                price = request.form.get("price")
+                update_product(prod_id, name, info, price)
+                flash('Вы обновили товар.', 'success')
+
+                # image
+                try:
+                    image = request.files['image']
+                    img_type = image.filename.split('.')[-1].lower()
+                    if img_type != '':
+                        types = ['jpeg', 'jpg', 'png']
+                        if img_type in types:
+                            resp_status, data = encrypt_id_with_no_digits(str(prod_id))
+                            if resp_status:
+                                product_id = data
+                                filename = ''
+                                directory = 'storage/products/'
+                                files = os.listdir(directory)
+                                for file in files:
+                                    if product_id == file.split('.')[0]:
+                                        filename = file
+                                        break
+                                if filename != '':
+                                    os.remove(directory + filename)
+                                filename = product_id + '.' + img_type
+                                image.save(directory + filename)
+                            else:
+                                flash('Не удалось обработать данные.', 'warning')
+                        else:
+                            flash('Недопустимый тип файла.', 'warning')
+                except Exception as ex:
+                    flash('Не удалось изменить изображение.', 'warning')
+                    logger.error(ex)
+            elif request.form['btn_products'] == 'delete_product':
+                prod_id = request.form.get("product_id")
+                try:
+                    resp_status, data = encrypt_id_with_no_digits(str(prod_id))
+                    if resp_status:
+                        product_id = data
+                        filename = ''
+                        directory = 'storage/products/'
+                        files = os.listdir(directory)
+                        for file in files:
+                            if product_id == file.split('.')[0]:
+                                filename = file
+                                break
+                        if filename != '':
+                            os.remove(directory + filename)
+                    else:
+                        flash('Не удалось обработать данные.', 'warning')
+                except Exception as ex:
+                    logger.error(ex)
+
+                delete_product(prod_id)
+                flash('Вы удалили товар.', 'success')
+        except Exception as ex:
+            logger.error(ex)
+            flash('Произошла какая-то ошибка', 'error')
+
+    orders = []
+    for i in get_orders().data:
+        # product
+        r = get_product_by_id(i.product)
+        product_name = str(i.product)
+        if r.success:
+            product_name = r.data.name
+        # client
+        r = get_user_by_id(i.client)
+        client_name = str(i.client)
+        if r.success:
+            client_name = f'{r.data.first_name} {r.data.last_name}'
+        orders.append({
+            "id": str(i.id),
+            "product": {"id": str(i.product),
+                        "name": product_name},
+            "client": {"id": str(i.client),
+                       "name": client_name},
+            "timestamp": i.timestamp.strftime("%d.%m.%Y %H:%M:%S"),
+            "status": i.status
+        })
+
+    products = []
+    for i in get_products().data:
+        product_status, product_id = encrypt_id_with_no_digits(str(i.id))
+        if product_status:
+            products.append({
+                'id': f'{product_id}',
+                'product_id': str(i.id),
+                'name': i.name,
+                'info': i.info,
+                'price': i.price
+            })
+
+    return render_template("admin/other/products.html", orders=orders, products=products)
